+++ lib/autodev/errors.rb
@@ -2,6 +2,7 @@
 
 class AutodevError < StandardError; end
 
+require_relative 'errors/authentication_error'
 require_relative 'errors/config_error'
 require_relative 'errors/git_error'
 require_relative 'errors/implementation_error'